old/requirements.txt
flask
psutil
py-cpuinfo